Refactor team feature tests (#1136)
